Fixed UserAccessLevelRepository delete method

+++ backend/src/main/java/pl/lodz/p/it/auctionsystem/mok/services/UserAccessLevelServiceImpl.java
@@ -64,6 +64,7 @@ public class UserAccessLevelServiceImpl implements UserAccessLevelService {
     
     @Override
     public void deleteUserAccessLevel(Long userAccessLevelId) {
-        userAccessLevelRepository.deleteById(userAccessLevelId);
+        if (userAccessLevelRepository.existsById(userAccessLevelId))
+            userAccessLevelRepository.deleteById(userAccessLevelId);
     }
 }
